Reverts "Temporarily disable new unit tests"
Mounting directories from the host into docker containers appears to not
work in travis. This commit changes build_test to use volume mounts
instead and thus needs to copy files in and out of the container instead
of simply accessing the disk.

I have also improved the handling of go modules on travis:
- fixed the cache of go module packages
- stopped travis from running go get during install