Experimental `pack run`
Introduces a `pack run` command which builds a container and then
immedatitly runs the newly built container.

A subset of the `pack build` args are supported, including:
- `--path`
- `--builder`
- `--run-image`

And `--port` is introduced as the local port the running container will
receive request for.

The run command internally leverages the build command's infrastructure.

The reponame is the md5 hash of the working directory.