Pass UID & GID through to lifecycle exporter
In converting to using exporter outside docker (to receive credentials
on mac), we inadvertently caused files to be owned by the uid & gid of
the user on the host machine, since they could not be otherwise without
super user.

This commit sets the user and gid to the runtime user during tar creation.