Move tests that don't require specific builders to their own suite
The current setup requires complicated workarounds when the output of
specific commands changes.

For example, we require the pack_previous_fixtures_overrides directory
and files such as inspect_X.Y.Z_builder_output.txt to change the desired
expectation when a previous version of pack is used to either
create a builder or inspect a builder.

This is cumbersome. We can reduce the number of tests that require
these workarounds by only applying the compatibility matrix for tests
that truly require a compatibility check. For example, we do not need to
re-evaluate the output of `pack suggest-builders` from a previous version
of pack.

Signed-off-by: Natalie Arellano <narellano@vmware.com>