Reset permissions on test directories
A previous commit had restricted permissions on directories we create/move during our testing. Overall that is a very helpful process, but in this particular case, it was causing some of our ensuing code and tests to fail, given that they expected to be able to write to those directories; further, it was in a test context, where there is limited room for exposure of files. Given that, I've reverted this aspect of the PR. If someone wants to do a slower and more thorough analysis of the permissions we create, in order to ensure that they are appropriate and the most limited they can be, that would be gold.

Signed-off-by: David Freilich <david.freilich@appsflyer.com>