Remove `heroku/buildpacks:18` from trusted and suggested builders list
The `heroku/buildpacks:18` builder is based upon Heroku-18 / Ubuntu 18.04,
which are due to enter their deprecation cycle In the coming April/May.

In addition, we currently only run `heroku/buildpacks:20` on the Heroku platform,
so whilst this builder does have some rudimentary CI coverage in our builder generation
repository (`heroku/pack-images`), the older stack is unused and barely tested in
practice. We've recently found a few developers who were inadvertently still using
it, since they'd set it as the default builder set some time ago.

Lastly, we expected to release a new `heroku/buildpacks:22`builder based on Ubuntu-22
next year, at which point we'd have three builders in the suggested builders list, which
seems excessive :-)

As such, the older stack has been removed from the trusted and suggested builders
list. This is a breaking change for anyone using that image, since they will need to
explicitly add it to their trusted builders list (so this will need to be in a new major
Pack release).

Signed-off-by: Ed Morley <501702+edmorley@users.noreply.github.com>