Set the Pack Homebrew formula version explicitly
Previously the Pack Homebrew formula did not set an explicit version,
meaning that Homebrew attempts to detect it from the package URL.

Unfortunately on ARM64, this results in the Pack version being detected
as `64`, due to the additional `arm64` suffix in the package URL.

This results in `pack upgrade` not detecting new pack CLI versions, so
pack is never upgraded.

Now, the pack version is set explicitly, avoiding the issues with
auto-detection.

In addition, `version_scheme` has to be set, since otherwise Homebrew
will not update the existing installations, as `64` would be treated as a
newer version. See:
https://docs.brew.sh/Formula-Cookbook#version-scheme-changes

Lastly, whilst fixing this, I noticed a bug in the GitHub Actions workflow,
whereby `tag_name` incorrectly had the `v` prefix removed before
being passed to the GitHub releases API. This codepath is presumably
unused most of the time (since it's only used when the workflow is
manually triggered), which is why this had gone unnoticed.

Fixes #buildpacks/homebrew-tap#20.

Signed-off-by: Ed Morley <501702+edmorley@users.noreply.github.com>