if lifecycle image is supplied, consider its apis
if we're injecting a custom lifecycle image into an untrusted builder,
when we select the platform api version we'll consider the intersection
between the builder, the lifecycle, and this instance of pack
when choosing which version of the api to speak.

because this is a rare case, the default is that there is no lifecycle
api list, which is treated as no constraints.

Signed-off-by: Joe Kimmel <jkimmel@vmware.com>